move : 페이지 컴포넌트에서 스텝 관리 하도록 변경
